Add #valid_takes to pawn and #opponent_piece? to Piece class
I added one valid_takes method for each player in the pawn class to
lower the abc of the method.

The #opponent_piece? method takes the board as an argument and coordinates
on the board to check whether or not the coordinates contain a piece that
would be an opponent that can be taken by the current piece.

This is primarily important for the pawn, who is the only piece in the
game that gets a special way of moving when taking a piece.