Add #allied_piece? method and split the move methods
allied_piece? checks if there is an allied piece on a possible move so it
can be excluded by the validation methods.

I also split the #valid_one_forward and #valid_double_forward methods into
black and white variants to reduce abd size and make the if statement
easier on the upcoming method that gets all the possible moves.