Refactor #move_piece and rename #pawn_moves
move_piece not checks if the coordinates are the coordinates of a legal
move.

pawn_moves has been renamed to possible_moves to make is more universal.
This will come into play later, when the actual piece selected by the
player cannot be known by me, the programmer,and will hopefully allow me
to avoid writing an if case for every possible piece.