Add #promote_pawn and fix a bug
The #promote_pawn_white and #promote_pawn_black methods return if the
pawn is not yet in the last row. If it is, it asks for input to decide
what piece to promote the pawn to and uses a switch statement to transform
the pawn.

The fixed bug was in #valid_one_forward, where I discovered that because
I was not checking to see if the pawn had an opponent in front of it, my
pawns could take the piece that was in front of them.