Add #handle_king
Just like #handle_pawn, #handle_king is called by #update_board if the
piece select by the player is the king. The method automatically returns
if the king has already moved, as that means the king can only move in
his usual pattern.

If the king has not moved yet, it checks if any castles would be legal
and performs them if the player has requested them.

It does so simply by moving the rook to its proper position, since the
king will be moved by the caller, #update_board.

After performing the castle, it changes the #has_moved instance variable
of the king to true so that the player cannot castle twice later on.