Add #check_for_checks
The method does what it says, it checks if the king is in a check and
modifies its new #in_check attribute to true if he is.

For now I have built two helper methods that check if the king is in a
check from the Rook, Bishop or Queen.