Add #filter_checked_positions and #check_mate?
The first is added to the King class and the 2nd to the Game class.

I have also made the filtering of allied pieces for the king into a
method to reduce abc size for #possible_moves and removed the puts
in the #check_from_(piece)? methods to avoid the command line being
spammed with 'check from piece on row, col' when #filter_checked_positions
simulates all the different scenarios.

For now, #check_mate? only checks if the array returned by #possible_moves
is empty, but I also need to add a way for it to check if the piece
giving the check cannot be taken by a piece allied to the king.