Add #moves_next_to_opponent_king
This method returns an array of king moves that would result in the king
being next to the enemy king.

It does so by simulating each possible king position and returning that
position if it would result in the king having another king on any of his
surrounding spaces.

This is used in the king's #possible_moves method to filter out those
positions.

I have also added an if clause to every piece type that removes a
position from the list of possible positions if that position contains
the enemy king, since pieces should not be able to take kings.