Refactor #choose_piece, #move_piece and #play
Turns out there are a ton of different ways to get a draw in chess and
a lot of them are so situational it would take days to even research all
of them.

Given that this game is played in the command line, and so will never
be played by two players who are not in the same room, I have simply
added the possibility for the player input to accept a string 'draw'
and end the game should that string be typed when the game asks for the
coordinates of a new piece.

This way the two players can decide on a draw whenever they wish or
end their game in a draw should they ever reach a position in which
winning is not legally possible.