Fix a bug with #different_move
Turns out that making a backup of the board doesn't work, because
making the backup equal to board.board means that it will update the
backup as well when the board gets updated.

I have refactored it to receive a backup of the old piece instead and
to use the coordinates selected by the player in order to manually bring
the board back to the state it was in before.