Refactor play and add #draw? to input_handler
Combined the two #puts into one and moved board.display_board to be only
at the start of the loop.

The effect is the same, but it is more efficient this way.

Making #draw? a part of InputHandler is more costly, but more logical
and it makes it easier to understand what #play is doing: gathering all
the other methods and combining them to make the game playable.