Expand #handle_pawn and refactor #update_board
Added an #en_passantable instance variable to the Pawn class that is set
to false by default and set to true by #handle_pawn when a pawn moves
two squares at once.

Also realized that #handle_pawn will have to do the updating of col/row
and of the board, just like handle_king, so I have placed it at the
start of #update_board with a return condition.