Remove comment from chess.rb
Reworking #opponent_piece? and #allied_piece? to only need the #player
value of the piece and the cell in question is possible, but it would
require a lot of testing to ensure that nothing breaks in the process.

I could do it without difficulty, but at this point I wish to move on
and learn new things. If anything, this has shown me the importance of
having good up-to-date tests for every class, as having those would have
allowed me to rewrite the two methods and be able to test them without
much trouble.